config changes
- update OLED display to use 2 lines
- update prox sensor to poll at 200hz
- add button callback (next to resset) to trigger move to obstacle
- add encoder globals (*_DIRECTIONAL) => forward is positive, back negative
- Extend servo turn mag by 1

+++ stm32/Core/Src/ir_adc.c
@@ -6,6 +6,7 @@
 // static uint16_t raw;
 
 #include "ir_adc.h"
+#include "stdlib.h"
 #include "math.h"
 
 static ADC_HandleTypeDef *HADC;
@@ -40,7 +41,7 @@ void _accumulate_sum(void)
     uint32_t fraction = AGGREGATE_SUM >> IR_ADC_SUM_SIZE_BITS;
 
     // ignore if out of range - periodic jitter from signal
-    if (abs(fraction - ADC_RAW_READOUT) > IR_ADC_REJECT_THRESHOLD)
+    if (abs((int)(fraction - ADC_RAW_READOUT)) > IR_ADC_REJECT_THRESHOLD)
     {
         return;
     }
